// src/app/modules/configuration/finance/models/late-fee-view.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * 
 */
export interface LateFeeView { 
    /**
     * Gets or Sets of Id
     */
    id: string;
    /**
     * 
     */
    delayFromDayCount: number;
    /**
     * 
     */
    delayToDayCount: number;
    /**
     * 
     */
    lateFeePenaltyTypeId: string;
    /**
     * 
     */
    penaltyAmount: number;
    /**
     * 
     */
    isPenaltyFromDueDate?: number;
}